only UserStatsService missing. Changed id to long/changed update methods/changed in services where IDao dao existed to DaoImp dao

+++ src/main/java/org/afdemp/wellness/services/ICartService.java
@@ -8,8 +8,8 @@ import org.afdemp.wellness.entities.Cart;
 public interface ICartService {
     
     public List<Cart> findAllCarts();
-    public Cart findById(int id);
+    public Cart findById(long id);
     public boolean save(Cart cart);
-    public boolean delete(int id);
+    public boolean delete(long id);
     public boolean update(Cart cart);
 }
